refactor: typescript-ize translation.js
